+++ pubtools/pulplib/_impl/model/__init__.py
@@ -9,6 +9,9 @@ from .repository import (
     FileSyncOptions,
     ContainerSyncOptions,
     YumSyncOptions,
+    Importer,
+    FileImporter,
+    YumImporter,
 )
 from .unit import (
     Unit,