+++ pubtools/pulplib/_impl/model/unit/erratum.py
@@ -1,8 +1,9 @@
 from frozenlist2 import frozenlist
 
-from .base import Unit, PulpObject, unit_type, schemaless_init
+from .base import Unit, PulpObject, unit_type
 
 from ..attr import pulp_attrib
+from ..common import schemaless_init
 from ... import compat_attr as attr
 from ..validate import (
     optional_bool,