+++ pubtools/pulplib/_impl/model/unit/rpm.py
@@ -1,9 +1,10 @@
 import datetime
 
 from pubtools.pulplib._impl.model.validate import optional_list_of
-from .base import Unit, unit_type, schemaless_init
+from .base import Unit, unit_type
 
 from ..attr import pulp_attrib
+from ..common import schemaless_init
 from ... import compat_attr as attr
 from ..convert import (
     frozenlist_or_none_converter,